Fix bugs
- Fix RegMulti to work properly
- Remove deprecated warning of CopyOrExpand
- In Set,%DestVar%,Permanent, if a value exists in interface, the value of interface will be changed instead of script.project.
- Fix crash if a plugin does not contain [Process] section
- RadioGroup's margin is properly rendered

+++ PEBakery.Tests/Core/Command/CommandArchiveTests.cs
@@ -225,42 +225,42 @@ namespace PEBakery.Tests.Core.Command
         [TestMethod]
         public void Archive_Expand()
         {
+            EngineState s = EngineTests.CreateEngineState();
             string rawCode; // $"Expand,\"%TestBench%\\CommandArchive\\{archiveFile}\",\"%TestBench%\\CommandArchive\\Decompress\""
 
             rawCode = $"Expand,\"%TestBench%\\CommandArchive\\{"ex1.cab"}\",\"%TestBench%\\CommandArchive\\Expand_0\"";
-            Expand_FileTemplate("ex1.cab", "ex1.jpg", 0, rawCode, ErrorCheck.Success);
+            Expand_FileTemplate(s, "ex1.cab", "ex1.jpg", 0, rawCode, ErrorCheck.Success);
 
             rawCode = $"Expand,\"%TestBench%\\CommandArchive\\{"ex2.cab"}\",\"%TestBench%\\CommandArchive\\Expand_1\",ex2.jpg";
-            Expand_FileTemplate("ex2.cab", "ex2.jpg", 1, rawCode, ErrorCheck.Success);
+            Expand_FileTemplate(s, "ex2.cab", "ex2.jpg", 1, rawCode, ErrorCheck.Success);
 
             rawCode = $"Expand,\"%TestBench%\\CommandArchive\\{"ex2.cab"}\",\"%TestBench%\\CommandArchive\\Expand_2\",ex1.jpg";
-            Expand_FileTemplate("ex2.cab", "ex1.jpg", 2, rawCode, ErrorCheck.Error);
+            Expand_FileTemplate(s, "ex2.cab", "ex1.jpg", 2, rawCode, ErrorCheck.Error);
 
             rawCode = $"Expand,\"%TestBench%\\CommandArchive\\{"ex3.jp_"}\",\"%TestBench%\\CommandArchive\\Expand_3\",ex3.jpg,PRESERVE";
-            Expand_FileTemplate("ex3.jp_", "ex3.jpg", 3, rawCode, ErrorCheck.Success, false, true);
+            Expand_FileTemplate(s, "ex3.jp_", "ex3.jpg", 3, rawCode, ErrorCheck.Success, false, true);
 
             rawCode = $"Expand,\"%TestBench%\\CommandArchive\\{"ex3.jp_"}\",\"%TestBench%\\CommandArchive\\Expand_4\",ex3.jpg,PRESERVE";
-            Expand_FileTemplate("ex3.jp_", "ex3.jpg", 4, rawCode, ErrorCheck.Warning, true, false);
+            Expand_FileTemplate(s, "ex3.jp_", "ex3.jpg", 4, rawCode, ErrorCheck.Warning, true, false);
 
             rawCode = $"Expand,\"%TestBench%\\CommandArchive\\{"ex3.jp_"}\",\"%TestBench%\\CommandArchive\\Expand_5\",ex3.jpg,PRESERVE,NOWARN";
-            Expand_FileTemplate("ex3.jp_", "ex3.jpg", 5, rawCode, ErrorCheck.Success, true, false);
+            Expand_FileTemplate(s, "ex3.jp_", "ex3.jpg", 5, rawCode, ErrorCheck.Success, true, false);
 
             rawCode = $"Expand,\"%TestBench%\\CommandArchive\\{"ex4.cab"}\",\"%TestBench%\\CommandArchive\\Expand_6\"";
-            Expand_DirTemplate("ex4.cab", "Cab", 6, rawCode, ErrorCheck.Success);
+            Expand_DirTemplate(s, "ex4.cab", "Cab", 6, rawCode, ErrorCheck.Success);
 
             rawCode = $"Expand,\"%TestBench%\\CommandArchive\\{"ex4.cab"}\",\"%TestBench%\\CommandArchive\\Expand_7\",ex3.jpg";
-            Expand_FileTemplate("ex4.cab", "ex3.jpg", 7, rawCode, ErrorCheck.Success);
+            Expand_FileTemplate(s, "ex4.cab", "ex3.jpg", 7, rawCode, ErrorCheck.Success);
 
             rawCode = $"Expand,\"%TestBench%\\CommandArchive\\{"ex4.cab"}\",\"%TestBench%\\CommandArchive\\Expand_8\",ex2.jpg,NOWARN";
-            Expand_FileTemplate("ex4.cab", "ex2.jpg", 8, rawCode, ErrorCheck.Success, true);
+            Expand_FileTemplate(s, "ex4.cab", "ex2.jpg", 8, rawCode, ErrorCheck.Success, true);
 
             rawCode = $"Expand,\"%TestBench%\\CommandArchive\\{"ex4.cab"}\",\"%TestBench%\\CommandArchive\\Expand_9\",ex2.jpg,NOWARN";
-            Expand_FileTemplate("ex4.cab", "ex2.jpg", 9, rawCode, ErrorCheck.Success, false);
+            Expand_FileTemplate(s, "ex4.cab", "ex2.jpg", 9, rawCode, ErrorCheck.Success, false);
         }
 
-        public void Expand_FileTemplate(string archiveFile, string compFile, int rev, string rawCode, ErrorCheck check, bool testPreserve = false, bool checkIfPreserve = true)
+        public void Expand_FileTemplate(EngineState s, string archiveFile, string compFile, int rev, string rawCode, ErrorCheck check, bool testPreserve = false, bool checkIfPreserve = true)
         {
-            EngineState s = EngineTests.CreateEngineState();
             string dirPath = StringEscaper.Preprocess(s, Path.Combine("%TestBench%", "CommandArchive"));
             string srcPath = Path.Combine(dirPath, "Cab", compFile);
             string destPath = Path.Combine(dirPath, $"Expand_{rev}", compFile);
@@ -275,8 +275,7 @@ namespace PEBakery.Tests.Core.Command
 
                 EngineTests.Eval(s, rawCode, CodeType.Expand, check);
 
-                if ((!testPreserve && File.Exists(destPath)) 
-                    || (testPreserve && checkIfPreserve))
+                if ((!testPreserve && File.Exists(destPath)) || (testPreserve && checkIfPreserve))
                 {
                     using (FileStream srcStream = new FileStream(srcPath, FileMode.Open, FileAccess.Read, FileShare.Read))
                     using (FileStream destStream = new FileStream(destPath, FileMode.Open, FileAccess.Read, FileShare.Read))
@@ -287,7 +286,6 @@ namespace PEBakery.Tests.Core.Command
                     }
                 }
                 
-
                 Console.WriteLine($"{archiveFile}, {rev} Success");
             }
             finally
@@ -296,9 +294,8 @@ namespace PEBakery.Tests.Core.Command
             }
         }
 
-        public void Expand_DirTemplate(string archiveFile, string compDir, int rev, string rawCode, ErrorCheck check, bool testPreserve = false, bool checkIfPreserve = true)
+        public void Expand_DirTemplate(EngineState s, string archiveFile, string compDir, int rev, string rawCode, ErrorCheck check, bool testPreserve = false, bool checkIfPreserve = true)
         { 
-            EngineState s = EngineTests.CreateEngineState();
             string dirPath = StringEscaper.Preprocess(s, Path.Combine("%TestBench%", "CommandArchive"));
             string srcPath = Path.Combine(dirPath, "Cab");
             string destPath = Path.Combine(dirPath, $"Expand_{rev}");
@@ -345,25 +342,29 @@ namespace PEBakery.Tests.Core.Command
         [TestMethod]
         public void Archive_CopyOrExpand()
         {
-            CopyOrExpand_1();
-            CopyOrExpand_2();
-            CopyOrExpand_3();
+            EngineState s = EngineTests.CreateEngineState();
+
+            CopyOrExpand_1(s);
+            CopyOrExpand_2(s);
+            CopyOrExpand_3(s);
         }
 
-        public void CopyOrExpand_1()
+        public void CopyOrExpand_1(EngineState s)
         {
-            EngineState s = EngineTests.CreateEngineState();
             string dirPath = StringEscaper.Preprocess(s, Path.Combine("%TestBench%", "CommandArchive"));
             string srcPath = Path.Combine(dirPath, "Cab", "ex3.jpg");
             string destDir = Path.Combine(dirPath, "CopyOrExpand_1");
             string destFile = Path.Combine(destDir, "ex3.jpg");
+
+            if (Directory.Exists(destDir))
+                Directory.Delete(destDir, true);
 
             try
             {
                 Directory.CreateDirectory(destDir);
 
                 string rawCode = "CopyOrExpand,\"%TestBench%\\CommandArchive\\ex3.jpg\",\"%TestBench%\\CommandArchive\\CopyOrExpand_1\"";
-                EngineTests.Eval(s, rawCode, CodeType.CopyOrExpand, ErrorCheck.Warning);
+                EngineTests.Eval(s, rawCode, CodeType.CopyOrExpand, ErrorCheck.Success);
 
                 Assert.IsTrue(Directory.Exists(destDir));
                 Assert.IsTrue(File.Exists(destFile));
@@ -378,24 +379,27 @@ namespace PEBakery.Tests.Core.Command
             }
             finally
             {
-                Directory.Delete(destDir, true);
+                if (Directory.Exists(destDir))
+                    Directory.Delete(destDir, true);
             }
         }
 
-        public void CopyOrExpand_2()
+        public void CopyOrExpand_2(EngineState s)
         {
-            EngineState s = EngineTests.CreateEngineState();
             string dirPath = StringEscaper.Preprocess(s, Path.Combine("%TestBench%", "CommandArchive"));
             string srcPath = Path.Combine(dirPath, "Cab", "ex3.jpg");
             string destDir = Path.Combine(dirPath, "CopyOrExpand_2");
             string destFile = Path.Combine(destDir, "change.jpg");
+
+            if (Directory.Exists(destDir))
+                Directory.Delete(destDir, true);
 
             try
             {
                 Directory.CreateDirectory(destDir);
 
                 string rawCode = "CopyOrExpand,\"%TestBench%\\CommandArchive\\ex3.jpg\",\"%TestBench%\\CommandArchive\\CopyOrExpand_2\\change.jpg\"";
-                EngineTests.Eval(s, rawCode, CodeType.CopyOrExpand, ErrorCheck.Warning);
+                EngineTests.Eval(s, rawCode, CodeType.CopyOrExpand, ErrorCheck.Success);
 
                 Assert.IsTrue(Directory.Exists(destDir));
                 Assert.IsTrue(File.Exists(destFile));
@@ -410,14 +414,15 @@ namespace PEBakery.Tests.Core.Command
             }
             finally
             {
-                Directory.Delete(destDir, true);
+                if (Directory.Exists(destDir))
+                    Directory.Delete(destDir, true);
             }
         }
 
-        public void CopyOrExpand_3()
+        public void CopyOrExpand_3(EngineState s)
         {
             string rawCode = "CopyOrExpand,\"%TestBench%\\CommandArchive\\ex5.jpg\",\"%TestBench%\\CommandArchive\\CopyOrExpand_3\"";
-            EngineTests.Eval(rawCode, CodeType.CopyOrExpand, ErrorCheck.Error);
+            EngineTests.Eval(s, rawCode, CodeType.CopyOrExpand, ErrorCheck.Error);
         }
         #endregion
